+++ includes/shared_bootstrap.php
@@ -34,6 +34,7 @@ $sharedFiles = [
     __DIR__ . '/../src/LandingPage/ResponsiveImageBuilder.php',
     __DIR__ . '/../src/Notification/TemplateVariableReplacer.php',
     __DIR__ . '/../src/Activity/ActivityQueryService.php',
+    __DIR__ . '/../src/Activity/ActivitySummaryCards.php',
     __DIR__ . '/../src/Lead/LeadCsvExportService.php',
     __DIR__ . '/../src/Agency/SubAgentCsvExportService.php',
     __DIR__ . '/../src/Agency/RecruitmentLinkCsvExportService.php',